tools/requirements.txt
mcp==1.27.0
fastmcp==3.2.1
ollama==0.6.1
pydantic==2.12.5
ddgs==9.13.0
psutil==7.2.2
numpy==2.4.4
PyYAML==6.0.3
llama-cpp-python==0.3.20
pymupdf==1.27.2.2
python-docx==1.2.0
ebooklib==0.20
beautifulsoup4==4.14.3
requests==2.32.3
tqdm==4.67.1
httpx==0.28.1
lxml==6.0.2
pathspec==0.12.1
